refactor: ButtonControl to Vue setup
