+++ source/win-main/file-manager/util/PopoverDirProps.vue
@@ -90,7 +90,7 @@ import formatDate from '@common/util/format-date'
 import localiseNumber from '@common/util/localise-number'
 import SelectControl from '@common/vue/form/elements/Select.vue'
 import SwitchControl from '@common/vue/form/elements/Switch.vue'
-import ButtonControl from '@common/vue/form/elements/Button.vue'
+import ButtonControl from '@common/vue/form/elements/ButtonControl.vue'
 import { trans } from '@common/i18n-renderer'
 import { type DirDescriptor, type MDFileDescriptor } from '@dts/common/fsal'
