+++ source/win-update/App.vue
@@ -90,7 +90,7 @@
  */
 
 import WindowChrome from '@common/vue/window/Chrome.vue'
-import ButtonControl from '@common/vue/form/elements/Button.vue'
+import ButtonControl from '@common/vue/form/elements/ButtonControl.vue'
 import ProgressControl from '@common/vue/form/elements/Progress.vue'
 import { trans } from '@common/i18n-renderer'
 import formatSize from '@common/util/format-size'
